moved new idempotent tests to own file
